FEAT: 자정에서 정오로 변경
